requirements.txt
# Core ML/DL Frameworks
torch>=2.0.0
torch-geometric>=2.3.0
torchvision>=0.15.0
tensorflow>=2.13.0  # Alternative to PyTorch

# Graph Processing
networkx>=3.0
dgl>=1.1.0  # Deep Graph Library
pygraphviz>=1.10

# Code Analysis & Parsing
tree-sitter>=0.20.0 
astroid>=2.15.0
pylint>=2.17.0
radon>=6.0.1  # Code metrics
lizard>=1.17.10  # Cyclomatic complexity

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Natural Language Processing
transformers>=4.30.0
sentencepiece>=0.1.99

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.5.0
tensorboard>=2.13.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Database
pymongo>=4.4.0
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
hypothesis>=6.82.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0
pre-commit>=3.3.0

# Security
bandit>=1.7.5
safety>=2.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
mkdocs>=1.5.0

# API Documentation
flasgger>=0.9.7
swagger-ui-bundle>=0.0.9

# Frontend (if using Python-based rendering)
jinja2>=3.1.2

# Performance Profiling
memory-profiler>=0.61.0
line-profiler>=4.0.0

# Deployment
docker>=6.1.0
docker-compose>=1.29.0

# Cloud SDKs (optional)
boto3>=1.28.0  # AWS
azure-storage-blob>=12.17.0  # Azure
google-cloud-storage>=2.10.0  # GCP

# Monitoring
prometheus-client>=0.17.0
sentry-sdk>=1.28.0

# Logging
loguru>=0.7.0
python-json-logger>=2.0.7

# Task Queue (optional)
celery>=5.3.0
redis>=4.6.0

# Graph Visualization
pydot>=1.4.2
graphviz>=0.20.1
pyvis>=0.3.2

# Static Analysis Tools
pyflakes>=3.0.1
prospector>=1.10.2
vulture>=2.7  # Dead code detection

# Type Checking
pydantic>=2.0.0
typing-extensions>=4.7.0

# Development Tools
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# Version Control
gitpython>=3.1.32

# CLI Tools
click>=8.1.0
rich>=13.4.0  # Beautiful terminal output
typer>=0.9.0

# Rate Limiting
ratelimit>=2.2.1

# Caching
redis>=4.6.0
diskcache>=5.6.0

# Serialization
pickle5>=0.0.12
msgpack>=1.0.5

# XML/HTML Parsing
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Regular Expressions (advanced)
regex>=2023.6.3

# Compression
zstandard>=0.21.0

# Progress Bars
alive-progress>=3.1.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Configuration Management
configparser>=6.0.0
toml>=0.10.2

# HTTP Clients
httpx>=0.24.0
aiohttp>=3.8.0

# Async Support
asyncio>=3.4.3
aiofiles>=23.1.0

# Code Coverage
coverage>=7.2.0

# Linting for Security
pyre-check>=0.9.18

# Graph Algorithms
graph-tool>=2.45  # Advanced graph analysis (optional, requires compilation)

# Metrics
prometheus-flask-exporter>=0.22.0

# Environment Management
virtualenv>=20.24.0
pipenv>=2023.7.0

# Build Tools
setuptools>=68.0.0
wheel>=0.40.0
build>=0.10.0

# Package Security Scanning
pip-audit>=2.6.0
